fix: moved PreviewToolbar
