+++ src/toolbar/PreviewTooltip.scss
